FAI-451: added loading status, refactored reducer
